backend/pyproject.toml
[project]
name = "cardiovision-rppg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mediapipe>=0.10.33",
    "numpy>=2.3.0",
    "open-rppg>=0.1.1",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.2",
    "scipy>=1.16.0",
    "scikit-learn>=1.8.0",
    "setuptools<81",
    "tqdm>=4.67.3",
    "websockets>=14.2",
]
